2-11-12 ========
Got Pagoda working on a live web server
http://www.lcjazzkatz.com/pagoda/

There were a few changes that had to be made

All paths in the  require() functions had to have the leading '/'
removed to make path relative instead of absolute.

ob_start() function was added to prevent 'Can not modify headers' error.
ob_start() write html to a buffer so it can write it after header has been
modifed. ob_end_flush() is also used to close the buffer. This is not
necessary on a localserver but it does not affect it.

The most intrustive change was setting the temp folder for session_start().
The 'session_save_path('/tmp')' was added to set the path on the webserver
but it will throw and error if the function fires on a local host. So
when running on a localhost this (session_save_path('/tmp')) must be
commented out.

Here is the code snippet
==============================
function __construct()
	{
		//disable on local
		//enable on web server
		session_save_path('/tmp');
		session_start();
	}
===========================================

Additional notes
================
My next goals (Alex) are to create the different css files for each user.
Also, Travis, I am working with the fade in transition so please leave them
enabled if possible for now so I can work with them.

Yojan will be creating Upload Class to upload documents and files
in Pagoda.